app/requirements.txt
aiohttp==3.9.0
PyJWT==2.8.0
SQLAlchemy==2.0.23
flask==3.0.0
flask-sqlalchemy==3.1.1
mysqlclient==2.2.0
psycopg2==2.9.2
pymysql==1.1.0